User Validation Otp done.

+++ service/generateOtp.js
@@ -2,11 +2,16 @@ const models = require("../models");
 const moment = require("moment");
 
 exports.generateOtp = async (Email) => {
+    const UserData = await models.userOtp.findOne({where : { Email: Email }});
+    if(UserData){  
+        if(UserData.IsValidated){
+        return {success : false , message : "User Already Validated." };
+        }
+    }
     await models.userOtp.destroy({ where: { Email: Email} });
-
     const Otp = Math.floor(1000 + Math.random() * 9000);
     const createdTime = new Date();
     const expiryTime = moment.utc(createdTime).add(15, "m");
-    await models.userOtp.create({ Email , Otp, createdTime, expiryTime, });
-    return Otp;
+    await models.userOtp.create({ Email , Otp, CreatedTime :createdTime , ExpiryTime : expiryTime, });
+    return {success : true , message : Otp}
 }
